Moved user form out of home so I can reuse it for new priv match (invites) and amended UserController to find user by email.
